# App/BackEnd/Conduit.API/src/Conduit.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.io"
    },
    "description" : "Conduit API",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "Conduit API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
  "paths" : {
...
    "/articles/{slug}/comments" : {
      "get" : {
        "description" : "Get the comments for an article. Auth is optional",
        "operationId" : "GetArticleComments",
        "parameters" : [ {
          "description" : "Slug of the article that you want to get comments for",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MultipleCommentsResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Get comments for an article",
        "tags" : [ "Comments" ]
      },
      "post" : {
        "description" : "Create a comment for an article. Auth is required",
        "operationId" : "CreateArticleComment",
        "parameters" : [ {
          "description" : "Slug of the article that you want to create a comment for",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCommentRequest"
              }
            }
          },
          "description" : "Comment you want to create",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleCommentResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Create a comment for an article",
        "tags" : [ "Comments" ],
        "x-codegen-request-body-name" : "comment"
      }
    },
    "/articles/{slug}/comments/{id}" : {
      "delete" : {
        "description" : "Delete a comment for an article. Auth is required",
        "operationId" : "DeleteArticleComment",
        "parameters" : [ {
          "description" : "Slug of the article that you want to delete a comment for",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "description" : "ID of the comment you want to delete",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Delete a comment for an article",
        "tags" : [ "Comments" ]
      }
    },
    "/articles/{slug}/favorite" : {
      "delete" : {
        "description" : "Unfavorite an article. Auth is required",
        "operationId" : "DeleteArticleFavorite",
        "parameters" : [ {
          "description" : "Slug of the article that you want to unfavorite",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleArticleResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Unfavorite an article",
        "tags" : [ "Favorites" ]
      },
      "post" : {
        "description" : "Favorite an article. Auth is required",
        "operationId" : "CreateArticleFavorite",
        "parameters" : [ {
          "description" : "Slug of the article that you want to favorite",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleArticleResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Favorite an article",
        "tags" : [ "Favorites" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Profile" : {
        "example" : {
          "image" : "image",
          "following" : true,
          "bio" : "bio",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "following" : {
            "type" : "boolean"
          }
        },
        "required" : [ "bio", "following", "image", "username" ],
        "type" : "object"
      },
      "Article" : {
        "example" : {
          "tagList" : [ "tagList", "tagList" ],
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "author" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          },
          "description" : "description",
          "title" : "title",
          "body" : "body",
          "favoritesCount" : 0,
          "slug" : "slug",
          "updatedAt" : "2000-01-23T04:56:07.000+00:00",
          "favorited" : true
        },
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "tagList" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "favorited" : {
            "type" : "boolean"
          },
          "favoritesCount" : {
            "type" : "integer"
          },
          "author" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "required" : [ "author", "body", "createdAt", "description", "favorited", "favoritesCount", "slug", "tagList", "title", "updatedAt" ],
        "type" : "object"
      },
      "SingleArticleResponse" : {
        "example" : {
          "article" : {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          }
        },
        "properties" : {
          "article" : {
            "$ref" : "#/components/schemas/Article"
          }
        },
        "required" : [ "article" ],
        "type" : "object"
      },
...
      "Comment" : {
        "example" : {
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "author" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          },
          "id" : 0,
          "body" : "body",
          "updatedAt" : "2000-01-23T04:56:07.000+00:00"
        },
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "author" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "required" : [ "author", "body", "createdAt", "id", "updatedAt" ],
        "type" : "object"
      },
      "SingleCommentResponse" : {
        "example" : {
          "comment" : {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          }
        },
        "properties" : {
          "comment" : {
            "$ref" : "#/components/schemas/Comment"
          }
        },
        "required" : [ "comment" ],
        "type" : "object"
      },
      "MultipleCommentsResponse" : {
        "example" : {
          "comments" : [ {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          }, {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          } ]
        },
        "properties" : {
          "comments" : {
            "items" : {
              "$ref" : "#/components/schemas/Comment"
            },
            "type" : "array"
          }
        },
        "required" : [ "comments" ],
        "type" : "object"
      },
      "NewComment" : {
        "example" : {
          "body" : "body"
        },
        "properties" : {
          "body" : {
            "type" : "string"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      },
      "NewCommentRequest" : {
        "example" : {
          "comment" : {
            "body" : "body"
          }
        },
        "properties" : {
          "comment" : {
            "$ref" : "#/components/schemas/NewComment"
          }
        },
        "required" : [ "comment" ],
        "type" : "object"
      },
...
      "GenericErrorModel" : {
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        },
        "required" : [ "errors" ],
        "type" : "object"
      },
      "GenericErrorModel_errors" : {
        "properties" : {
          "body" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "Token" : {
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}